feat: refactor datanode keys
