+++ src/meta-srv/src/selector/weight_compute.rs
@@ -17,7 +17,7 @@ use std::collections::HashMap;
 use common_meta::peer::Peer;
 use itertools::{Itertools, MinMaxResult};
 
-use crate::key::{StatKey, StatValue};
+use crate::key::{DatanodeStatKey, DatanodeStatValue};
 use crate::selector::weighted_choose::WeightedItem;
 
 /// The [`WeightCompute`] trait is used to compute the weight array by heartbeats.
@@ -37,9 +37,12 @@ pub trait WeightCompute: Send + Sync {
 pub struct RegionNumsBasedWeightCompute;
 
 impl WeightCompute for RegionNumsBasedWeightCompute {
-    type Source = HashMap<StatKey, StatValue>;
+    type Source = HashMap<DatanodeStatKey, DatanodeStatValue>;
 
-    fn compute(&self, stat_kvs: &HashMap<StatKey, StatValue>) -> Vec<WeightedItem<Peer>> {
+    fn compute(
+        &self,
+        stat_kvs: &HashMap<DatanodeStatKey, DatanodeStatValue>,
+    ) -> Vec<WeightedItem<Peer>> {
         let mut region_nums = Vec::with_capacity(stat_kvs.len());
         let mut peers = Vec::with_capacity(stat_kvs.len());
 
@@ -98,32 +101,32 @@ mod tests {
 
     use super::{RegionNumsBasedWeightCompute, WeightCompute};
     use crate::handler::node_stat::{RegionStat, Stat};
-    use crate::key::{StatKey, StatValue};
+    use crate::key::{DatanodeStatKey, DatanodeStatValue};
 
     #[test]
     fn test_weight_compute() {
-        let mut stat_kvs: HashMap<StatKey, StatValue> = HashMap::default();
-        let stat_key = StatKey {
+        let mut stat_kvs: HashMap<DatanodeStatKey, DatanodeStatValue> = HashMap::default();
+        let stat_key = DatanodeStatKey {
             cluster_id: 1,
             node_id: 1,
         };
-        let stat_val = StatValue {
+        let stat_val = DatanodeStatValue {
             stats: vec![mock_stat_1()],
         };
         stat_kvs.insert(stat_key, stat_val);
-        let stat_key = StatKey {
+        let stat_key = DatanodeStatKey {
             cluster_id: 1,
             node_id: 2,
         };
-        let stat_val = StatValue {
+        let stat_val = DatanodeStatValue {
             stats: vec![mock_stat_2()],
         };
         stat_kvs.insert(stat_key, stat_val);
-        let stat_key = StatKey {
+        let stat_key = DatanodeStatKey {
             cluster_id: 1,
             node_id: 3,
         };
-        let stat_val = StatValue {
+        let stat_val = DatanodeStatValue {
             stats: vec![mock_stat_3()],
         };
         stat_kvs.insert(stat_key, stat_val);
